+++ _site/assets/CheckboxGroup-B86SszLO.js
@@ -1,0 +1,105 @@
+import{At as e,Bn as t,Fn as n,Fr as r,Hn as i,It as a,Jt as o,Ln as s,Lt as c,Mr as l,Nt as u,Pn as d,Rn as f,Rr as p,Tt as m,Vn as h,Xn as g,Yt as _,Zt as v,_t as y,er as b,ir as x,jt as S,nr as C,or as w,rn as T,sr as E,tn as D,tr as O,ur as k,vr as A,vt as j,yr as M,yt as N}from"./opensearch-xDqIHoJa.js";import{r as P}from"./Suffix-BnZ9laDa.js";import{c as F}from"./Card-Bf4RhZD7.js";import{t as I}from"./misc-v7XGxIq4.js";import{c as L}from"./index-D40mbV63.js";var R=()=>(()=>{let e=o(`75be776d8875fa17`);return e[0]||=O(`svg`,{viewBox:`0 0 64 64`,class:`check-icon`},[O(`path`,{d:`M50.42,16.76L22.34,39.45l-8.1-11.46c-1.12-1.58-3.3-1.96-4.88-0.84c-1.58,1.12-1.95,3.3-0.84,4.88l10.26,14.51  c0.56,0.79,1.42,1.31,2.38,1.45c0.16,0.02,0.32,0.03,0.48,0.03c0.8,0,1.57-0.27,2.2-0.78l30.99-25.03c1.5-1.21,1.74-3.42,0.52-4.92  C54.13,15.78,51.93,15.55,50.42,16.76z`})],-1)})(),z=()=>(()=>{let e=o(`c6eed899356c8404`);return e[0]||=O(`svg`,{viewBox:`0 0 100 100`,class:`line-icon`},[O(`path`,{d:`M80.2,55.5H21.4c-2.8,0-5.1-2.5-5.1-5.5l0,0c0-3,2.3-5.5,5.1-5.5h58.7c2.8,0,5.1,2.5,5.1,5.5l0,0C85.2,53.1,82.9,55.5,80.2,55.5z`})],-1)})(),B=d([n(`checkbox`,`
+ font-size: var(--n-font-size);
+ outline: none;
+ cursor: pointer;
+ display: inline-flex;
+ flex-wrap: nowrap;
+ align-items: flex-start;
+ word-break: break-word;
+ line-height: var(--n-size);
+ --n-merged-color-table: var(--n-color-table);
+ `,[f(`show-label`,`line-height: var(--n-label-line-height);`),d(`&:hover`,[n(`checkbox-box`,[s(`border`,`border: var(--n-border-checked);`)])]),d(`&:focus:not(:active)`,[n(`checkbox-box`,[s(`border`,`
+ border: var(--n-border-focus);
+ box-shadow: var(--n-box-shadow-focus);
+ `)])]),f(`inside-table`,[n(`checkbox-box`,`
+ background-color: var(--n-merged-color-table);
+ `)]),f(`checked`,[n(`checkbox-box`,`
+ background-color: var(--n-color-checked);
+ `,[n(`checkbox-icon`,[d(`.check-icon`,`
+ opacity: 1;
+ transform: scale(1);
+ `)])])]),f(`indeterminate`,[n(`checkbox-box`,[n(`checkbox-icon`,[d(`.check-icon`,`
+ opacity: 0;
+ transform: scale(.5);
+ `),d(`.line-icon`,`
+ opacity: 1;
+ transform: scale(1);
+ `)])])]),f(`checked, indeterminate`,[d(`&:focus:not(:active)`,[n(`checkbox-box`,[s(`border`,`
+ border: var(--n-border-checked);
+ box-shadow: var(--n-box-shadow-focus);
+ `)])]),n(`checkbox-box`,`
+ background-color: var(--n-color-checked);
+ border-left: 0;
+ border-top: 0;
+ `,[s(`border`,{border:`var(--n-border-checked)`})])]),f(`disabled`,{cursor:`not-allowed`},[f(`checked`,[n(`checkbox-box`,`
+ background-color: var(--n-color-disabled-checked);
+ `,[s(`border`,{border:`var(--n-border-disabled-checked)`}),n(`checkbox-icon`,[d(`.check-icon, .line-icon`,{fill:`var(--n-check-mark-color-disabled-checked)`})])])]),n(`checkbox-box`,`
+ background-color: var(--n-color-disabled);
+ `,[s(`border`,`
+ border: var(--n-border-disabled);
+ `),n(`checkbox-icon`,[d(`.check-icon, .line-icon`,`
+ fill: var(--n-check-mark-color-disabled);
+ `)])]),s(`label`,`
+ color: var(--n-text-color-disabled);
+ `)]),n(`checkbox-box-wrapper`,`
+ position: relative;
+ width: var(--n-size);
+ flex-shrink: 0;
+ flex-grow: 0;
+ user-select: none;
+ -webkit-user-select: none;
+ `),n(`checkbox-box`,`
+ position: absolute;
+ left: 0;
+ top: 50%;
+ transform: translateY(-50%);
+ height: var(--n-size);
+ width: var(--n-size);
+ display: inline-block;
+ box-sizing: border-box;
+ border-radius: var(--n-border-radius);
+ background-color: var(--n-color);
+ transition: background-color 0.3s var(--n-bezier);
+ `,[s(`border`,`
+ transition:
+ border-color .3s var(--n-bezier),
+ box-shadow .3s var(--n-bezier);
+ border-radius: inherit;
+ position: absolute;
+ left: 0;
+ right: 0;
+ top: 0;
+ bottom: 0;
+ border: var(--n-border);
+ `),n(`checkbox-icon`,`
+ display: flex;
+ align-items: center;
+ justify-content: center;
+ position: absolute;
+ left: 1px;
+ right: 1px;
+ top: 1px;
+ bottom: 1px;
+ `,[d(`.check-icon, .line-icon`,`
+ width: 100%;
+ fill: var(--n-check-mark-color);
+ opacity: 0;
+ transform: scale(0.5);
+ transform-origin: center;
+ transition:
+ fill 0.3s var(--n-bezier),
+ transform 0.3s var(--n-bezier),
+ opacity 0.3s var(--n-bezier),
+ border-color 0.3s var(--n-bezier);
+ `),y({left:`1px`,top:`1px`})])]),s(`label`,`
+ color: var(--n-text-color);
+ transition: color .3s var(--n-bezier);
+ user-select: none;
+ -webkit-user-select: none;
+ padding: var(--n-label-padding);
+ font-weight: var(--n-label-font-weight);
+ `,[d(`&:empty`,{display:`none`})])]),h(n(`checkbox`,`
+ --n-merged-color-table: var(--n-color-table-modal);
+ `)),i(n(`checkbox`,`
+ --n-merged-color-table: var(--n-color-table-popover);
+ `))]),V=[`id`],H=[`tabindex`,`aria-checked`,`aria-labelledby`,`onKeyup`,`onKeydown`,`onClick`],U={...a.props,size:String,checked:{type:[Boolean,String,Number],default:void 0},defaultChecked:{type:[Boolean,String,Number],default:!1},value:[String,Number],disabled:{type:Boolean,default:void 0},indeterminate:Boolean,label:String,focusable:{type:Boolean,default:!0},checkedValue:{type:[Boolean,String,Number],default:!0},uncheckedValue:{type:[Boolean,String,Number],default:!1},"onUpdate:checked":[Function,Array],onUpdateChecked:[Function,Array],privateInsideTable:Boolean,onChange:[Function,Array]},W=E({name:`Checkbox`,props:U,setup(e){let n=k(G,null),i=l(null),{mergedClsPrefixRef:o,inlineThemeDisabled:s,mergedRtlRef:d,mergedComponentPropsRef:f}=D(e),p=l(e.defaultChecked),h=r(e,`checked`),g=P(h,p),_=u(()=>{if(n){let t=n.valueSetRef.value;return t&&e.value!==void 0?t.has(e.value):!1}return g.value===e.checkedValue}),v=N(e,{mergedSize(t){let{size:r}=e;if(r!==void 0)return r;if(n){let{value:e}=n.mergedSizeRef;if(e!==void 0)return e}if(t){let{mergedSize:e}=t;if(e!==void 0)return e.value}return f?.value?.Checkbox?.size||`medium`},mergedDisabled(t){let{disabled:r}=e;if(r!==void 0)return r;if(n){if(n.disabledRef.value)return!0;let{maxRef:{value:e},checkedCountRef:t}=n;if(e!==void 0&&t.value>=e&&!_.value)return!0;let{minRef:{value:r}}=n;if(r!==void 0&&t.value<=r&&_.value)return!0}return t?t.disabled.value:!1}}),{mergedDisabledRef:y,mergedSizeRef:x}=v,C=a(`Checkbox`,`-checkbox`,B,L,e,o);function w(t){if(n&&e.value!==void 0)n.toggleCheckbox(!_.value,e.value);else{let{onChange:n,"onUpdate:checked":r,onUpdateChecked:i}=e,{nTriggerFormInput:a,nTriggerFormChange:o}=v,s=_.value?e.uncheckedValue:e.checkedValue;r&&S(r,s,t),i&&S(i,s,t),n&&S(n,s,t),a(),o(),p.value=s}}function T(e){y.value||w(e)}function E(e){if(!y.value)switch(e.key){case` `:case`Enter`:w(e)}}function O(e){e.key===` `&&e.preventDefault()}let A={focus:()=>{i.value?.focus()},blur:()=>{i.value?.blur()}},j=m(`Checkbox`,d,o),M=b(()=>{let{value:e}=x,{common:{cubicBezierEaseInOut:n},self:{borderRadius:r,color:i,colorChecked:a,colorDisabled:o,colorTableHeader:s,colorTableHeaderModal:c,colorTableHeaderPopover:l,checkMarkColor:u,checkMarkColorDisabled:d,border:f,borderFocus:p,borderDisabled:m,borderChecked:h,boxShadowFocus:g,textColor:_,textColorDisabled:v,checkMarkColorDisabledChecked:y,colorDisabledChecked:b,borderDisabledChecked:S,labelPadding:w,labelLineHeight:T,labelFontWeight:E,[t(`fontSize`,e)]:D,[t(`size`,e)]:O}}=C.value;return{"--n-label-line-height":T,"--n-label-font-weight":E,"--n-size":O,"--n-bezier":n,"--n-border-radius":r,"--n-border":f,"--n-border-checked":h,"--n-border-focus":p,"--n-border-disabled":m,"--n-border-disabled-checked":S,"--n-box-shadow-focus":g,"--n-color":i,"--n-color-checked":a,"--n-color-table":s,"--n-color-table-modal":c,"--n-color-table-popover":l,"--n-color-disabled":o,"--n-color-disabled-checked":b,"--n-text-color":_,"--n-text-color-disabled":v,"--n-check-mark-color":u,"--n-check-mark-color-disabled":d,"--n-check-mark-color-disabled-checked":y,"--n-font-size":D,"--n-label-padding":w}}),F=s?c(`checkbox`,b(()=>x.value[0]),M,e):void 0;return Object.assign(v,A,{rtlEnabled:j,selfRef:i,mergedClsPrefix:o,mergedDisabled:y,renderedChecked:_,mergedTheme:C,labelId:I(),handleClick:T,handleKeyUp:E,handleKeyDown:O,cssVars:s?void 0:M,themeClass:F?.themeClass,onRender:F?.onRender})},render(){let{$slots:t,renderedChecked:n,mergedDisabled:r,indeterminate:i,privateInsideTable:a,cssVars:s,labelId:c,label:l,mergedClsPrefix:u,focusable:d,handleKeyUp:f,handleKeyDown:m,handleClick:h}=this;this.onRender?.();let g=e(t.default,e=>l||e?(A(),x(`span`,{key:1,class:_(`${u}-checkbox__label`),id:c},[v(()=>l||e)],10,V)):null);return(()=>{let e=o(`70be6e74cd27cb50`);return A(),x(`div`,{ref:`selfRef`,class:_([`${u}-checkbox`,this.themeClass,this.rtlEnabled&&`${u}-checkbox--rtl`,n&&`${u}-checkbox--checked`,r&&`${u}-checkbox--disabled`,i&&`${u}-checkbox--indeterminate`,a&&`${u}-checkbox--inside-table`,g&&`${u}-checkbox--show-label`]),tabindex:r||!d?void 0:0,role:`checkbox`,"aria-checked":i?`mixed`:n,"aria-labelledby":c,style:p(s),onKeyup:f,onKeydown:m,onClick:h,onMousedown:e[0]||=()=>{F(`selectstart`,window,e=>{e.preventDefault()},{once:!0})}},[O(`div`,{class:_(`${u}-checkbox-box-wrapper`)},[e[1]||=v(`\xA0`,-1),O(`div`,{class:_(`${u}-checkbox-box`)},[w(j,null,{default:()=>this.indeterminate?(A(),x(`div`,{key:`indeterminate`,class:_(`${u}-checkbox-icon`)},[v(()=>z())],2)):(A(),x(`div`,{key:`check`,class:_(`${u}-checkbox-icon`)},[v(()=>R())],2))},1024),O(`div`,{class:_(`${u}-checkbox-box__border`)},null,2)],2)],2),v(()=>g)],46,H)})()}}),G=T(`n-checkbox-group`);E({name:`CheckboxGroup`,props:{min:Number,max:Number,size:String,options:Array,labelField:{type:String,default:`label`},valueField:{type:String,default:`value`},value:Array,defaultValue:{type:Array,default:null},disabled:{type:Boolean,default:void 0},"onUpdate:value":[Function,Array],onUpdateValue:[Function,Array],onChange:[Function,Array]},setup(e){let{mergedClsPrefixRef:t}=D(e),n=N(e),{mergedSizeRef:i,mergedDisabledRef:a}=n,o=l(e.defaultValue),s=b(()=>e.value),c=P(s,o),u=b(()=>c.value?.length||0),d=b(()=>Array.isArray(c.value)?new Set(c.value):new Set);function f(t,r){let{nTriggerFormInput:i,nTriggerFormChange:a}=n,{onChange:s,"onUpdate:value":l,onUpdateValue:u}=e;if(Array.isArray(c.value)){let e=Array.from(c.value),n=e.findIndex(e=>e===r);t?~n||(e.push(r),u&&S(u,e,{actionType:`check`,value:r}),l&&S(l,e,{actionType:`check`,value:r}),i(),a(),o.value=e,s&&S(s,e)):~n&&(e.splice(n,1),u&&S(u,e,{actionType:`uncheck`,value:r}),l&&S(l,e,{actionType:`uncheck`,value:r}),s&&S(s,e),o.value=e,i(),a())}else t?(u&&S(u,[r],{actionType:`check`,value:r}),l&&S(l,[r],{actionType:`check`,value:r}),s&&S(s,[r]),o.value=[r],i(),a()):(u&&S(u,[],{actionType:`uncheck`,value:r}),l&&S(l,[],{actionType:`uncheck`,value:r}),s&&S(s,[]),o.value=[],i(),a())}return M(G,{checkedCountRef:u,maxRef:r(e,`max`),minRef:r(e,`min`),valueSetRef:d,disabledRef:a,mergedSizeRef:i,toggleCheckbox:f}),{mergedClsPrefix:t}},render(){let{options:e,labelField:t,valueField:n}=this.$props;return A(),x(`div`,{class:_(`${this.mergedClsPrefix}-checkbox-group`),role:`group`},[e?(A(),x(g,{key:0},[v(()=>e.map(e=>{let r=e[n];return A(),C(W,{key:r,value:r,disabled:e.disabled,label:e[t]},null,8,[`value`,`disabled`,`label`])}))],64)):(A(),x(g,{key:1},[v(()=>this.$slots.default?.())],64))],2)}});export{W as t};