+++ _site/assets/Input-DQWrCQXl.js
@@ -1,0 +1,234 @@
+import{At as e,Bn as t,Er as n,Fn as r,Fr as i,It as a,Jt as o,Kt as s,Ln as c,Lt as ee,Mr as l,Nt as u,Ot as d,Pn as f,Pt as p,Qt as te,Rn as m,Rr as h,Tr as g,Tt as ne,Xn as _,Yt as v,Zt as y,_r as re,cr as ie,er as b,fr as x,ht as ae,ir as S,jt as C,kt as w,nr as T,pr as E,rn as D,sr as O,tn as oe,tr as k,ur as A,vr as j,yr as se,yt as ce,zn as M}from"./opensearch-xDqIHoJa.js";import{i as le,n as N,r as ue,t as P}from"./Suffix-BnZ9laDa.js";import{c as F,i as I,n as L,s as R}from"./Card-Bf4RhZD7.js";import{u as de}from"./index-D40mbV63.js";var z=O({name:`Eye`,render(){return(()=>{let e=o(`ae479a1970012861`);return e[0]||=k(`svg`,{xmlns:`http://www.w3.org/2000/svg`,viewBox:`0 0 512 512`},[k(`path`,{d:`M255.66 112c-77.94 0-157.89 45.11-220.83 135.33a16 16 0 0 0-.27 17.77C82.92 340.8 161.8 400 255.66 400c92.84 0 173.34-59.38 221.79-135.25a16.14 16.14 0 0 0 0-17.47C428.89 172.28 347.8 112 255.66 112z`,fill:`none`,stroke:`currentColor`,"stroke-linecap":`round`,"stroke-linejoin":`round`,"stroke-width":`32`}),k(`circle`,{cx:`256`,cy:`256`,r:`80`,fill:`none`,stroke:`currentColor`,"stroke-miterlimit":`10`,"stroke-width":`32`})],-1)})()}}),B=O({name:`EyeOff`,render(){return(()=>{let e=o(`2c06203b450ce879`);return e[0]||=k(`svg`,{xmlns:`http://www.w3.org/2000/svg`,viewBox:`0 0 512 512`},[k(`path`,{d:`M432 448a15.92 15.92 0 0 1-11.31-4.69l-352-352a16 16 0 0 1 22.62-22.62l352 352A16 16 0 0 1 432 448z`,fill:`currentColor`}),k(`path`,{d:`M255.66 384c-41.49 0-81.5-12.28-118.92-36.5c-34.07-22-64.74-53.51-88.7-91v-.08c19.94-28.57 41.78-52.73 65.24-72.21a2 2 0 0 0 .14-2.94L93.5 161.38a2 2 0 0 0-2.71-.12c-24.92 21-48.05 46.76-69.08 76.92a31.92 31.92 0 0 0-.64 35.54c26.41 41.33 60.4 76.14 98.28 100.65C162 402 207.9 416 255.66 416a239.13 239.13 0 0 0 75.8-12.58a2 2 0 0 0 .77-3.31l-21.58-21.58a4 4 0 0 0-3.83-1a204.8 204.8 0 0 1-51.16 6.47z`,fill:`currentColor`}),k(`path`,{d:`M490.84 238.6c-26.46-40.92-60.79-75.68-99.27-100.53C349 110.55 302 96 255.66 96a227.34 227.34 0 0 0-74.89 12.83a2 2 0 0 0-.75 3.31l21.55 21.55a4 4 0 0 0 3.88 1a192.82 192.82 0 0 1 50.21-6.69c40.69 0 80.58 12.43 118.55 37c34.71 22.4 65.74 53.88 89.76 91a.13.13 0 0 1 0 .16a310.72 310.72 0 0 1-64.12 72.73a2 2 0 0 0-.15 2.95l19.9 19.89a2 2 0 0 0 2.7.13a343.49 343.49 0 0 0 68.64-78.48a32.2 32.2 0 0 0-.1-34.78z`,fill:`currentColor`}),k(`path`,{d:`M256 160a95.88 95.88 0 0 0-21.37 2.4a2 2 0 0 0-1 3.38l112.59 112.56a2 2 0 0 0 3.38-1A96 96 0 0 0 256 160z`,fill:`currentColor`}),k(`path`,{d:`M165.78 233.66a2 2 0 0 0-3.38 1a96 96 0 0 0 115 115a2 2 0 0 0 1-3.38z`,fill:`currentColor`})],-1)})()}}),V=D(`n-input`),fe=r(`input`,`
+ max-width: 100%;
+ cursor: text;
+ line-height: 1.5;
+ z-index: auto;
+ outline: none;
+ box-sizing: border-box;
+ position: relative;
+ display: inline-flex;
+ border-radius: var(--n-border-radius);
+ background-color: var(--n-color);
+ transition: background-color .3s var(--n-bezier);
+ font-size: var(--n-font-size);
+ font-weight: var(--n-font-weight);
+ --n-padding-vertical: calc((var(--n-height) - 1.5 * var(--n-font-size)) / 2);
+`,[c(`input, textarea`,`
+ overflow: hidden;
+ flex-grow: 1;
+ position: relative;
+ `),c(`input-el, textarea-el, input-mirror, textarea-mirror, separator, placeholder`,`
+ box-sizing: border-box;
+ font-size: inherit;
+ line-height: 1.5;
+ font-family: inherit;
+ border: none;
+ outline: none;
+ background-color: #0000;
+ text-align: inherit;
+ transition:
+ -webkit-text-fill-color .3s var(--n-bezier),
+ caret-color .3s var(--n-bezier),
+ color .3s var(--n-bezier),
+ text-decoration-color .3s var(--n-bezier);
+ `),c(`input-el, textarea-el`,`
+ -webkit-appearance: none;
+ scrollbar-width: none;
+ width: 100%;
+ min-width: 0;
+ text-decoration-color: var(--n-text-decoration-color);
+ color: var(--n-text-color);
+ caret-color: var(--n-caret-color);
+ background-color: transparent;
+ `,[f(`&::-webkit-scrollbar, &::-webkit-scrollbar-track-piece, &::-webkit-scrollbar-thumb`,`
+ width: 0;
+ height: 0;
+ display: none;
+ `),f(`&::placeholder`,`
+ color: #0000;
+ -webkit-text-fill-color: transparent !important;
+ `),f(`&:-webkit-autofill ~`,[c(`placeholder`,`display: none;`)])]),m(`round`,[M(`textarea`,`border-radius: calc(var(--n-height) / 2);`)]),c(`placeholder`,`
+ pointer-events: none;
+ position: absolute;
+ left: 0;
+ right: 0;
+ top: 0;
+ bottom: 0;
+ overflow: hidden;
+ color: var(--n-placeholder-color);
+ `,[f(`span`,`
+ width: 100%;
+ display: inline-block;
+ `)]),m(`textarea`,[c(`placeholder`,`overflow: visible;`)]),M(`autosize`,`width: 100%;`),m(`autosize`,[c(`textarea-el, input-el`,`
+ position: absolute;
+ top: 0;
+ left: 0;
+ height: 100%;
+ `)]),r(`input-wrapper`,`
+ overflow: hidden;
+ display: inline-flex;
+ flex-grow: 1;
+ position: relative;
+ padding-left: var(--n-padding-left);
+ padding-right: var(--n-padding-right);
+ `),c(`input-mirror`,`
+ padding: 0;
+ height: var(--n-height);
+ line-height: var(--n-height);
+ overflow: hidden;
+ visibility: hidden;
+ position: static;
+ white-space: pre;
+ pointer-events: none;
+ `),c(`input-el`,`
+ padding: 0;
+ height: var(--n-height);
+ line-height: var(--n-height);
+ `,[f(`&[type=password]::-ms-reveal`,`display: none;`),f(`+`,[c(`placeholder`,`
+ display: flex;
+ align-items: center; 
+ `)])]),M(`textarea`,[c(`placeholder`,`white-space: nowrap;`)]),c(`eye`,`
+ display: flex;
+ align-items: center;
+ justify-content: center;
+ transition: color .3s var(--n-bezier);
+ `),m(`textarea`,`width: 100%;`,[r(`input-word-count`,`
+ position: absolute;
+ right: var(--n-padding-right);
+ bottom: var(--n-padding-vertical);
+ `),m(`resizable`,[r(`input-wrapper`,`
+ resize: vertical;
+ min-height: var(--n-height);
+ `)]),c(`textarea-el, textarea-mirror, placeholder`,`
+ height: 100%;
+ padding-left: 0;
+ padding-right: 0;
+ padding-top: var(--n-padding-vertical);
+ padding-bottom: var(--n-padding-vertical);
+ word-break: break-word;
+ display: inline-block;
+ vertical-align: bottom;
+ box-sizing: border-box;
+ line-height: var(--n-line-height-textarea);
+ margin: 0;
+ resize: none;
+ white-space: pre-wrap;
+ scroll-padding-block-end: var(--n-padding-vertical);
+ `),c(`textarea-mirror`,`
+ width: 100%;
+ pointer-events: none;
+ overflow: hidden;
+ visibility: hidden;
+ position: static;
+ white-space: pre-wrap;
+ overflow-wrap: break-word;
+ `)]),m(`pair`,[c(`input-el, placeholder`,`text-align: center;`),c(`separator`,`
+ display: flex;
+ align-items: center;
+ transition: color .3s var(--n-bezier);
+ color: var(--n-text-color);
+ white-space: nowrap;
+ `,[r(`icon`,`
+ color: var(--n-icon-color);
+ `),r(`base-icon`,`
+ color: var(--n-icon-color);
+ `)])]),m(`disabled`,`
+ cursor: not-allowed;
+ background-color: var(--n-color-disabled);
+ `,[c(`border`,`border: var(--n-border-disabled);`),c(`input-el, textarea-el`,`
+ cursor: not-allowed;
+ color: var(--n-text-color-disabled);
+ text-decoration-color: var(--n-text-color-disabled);
+ `),c(`placeholder`,`color: var(--n-placeholder-color-disabled);`),c(`separator`,`color: var(--n-text-color-disabled);`,[r(`icon`,`
+ color: var(--n-icon-color-disabled);
+ `),r(`base-icon`,`
+ color: var(--n-icon-color-disabled);
+ `)]),r(`input-word-count`,`
+ color: var(--n-count-text-color-disabled);
+ `),c(`suffix, prefix`,`color: var(--n-text-color-disabled);`,[r(`icon`,`
+ color: var(--n-icon-color-disabled);
+ `),r(`internal-icon`,`
+ color: var(--n-icon-color-disabled);
+ `)])]),M(`disabled`,[c(`eye`,`
+ color: var(--n-icon-color);
+ cursor: pointer;
+ `,[f(`&:hover`,`
+ color: var(--n-icon-color-hover);
+ `),f(`&:active`,`
+ color: var(--n-icon-color-pressed);
+ `)]),f(`&:hover`,`background-color: var(--n-color-hover);`,[c(`state-border`,`border: var(--n-border-hover);`)]),m(`focus`,`background-color: var(--n-color-focus);`,[c(`state-border`,`
+ border: var(--n-border-focus);
+ box-shadow: var(--n-box-shadow-focus);
+ `)])]),c(`border, state-border`,`
+ box-sizing: border-box;
+ position: absolute;
+ left: 0;
+ right: 0;
+ top: 0;
+ bottom: 0;
+ pointer-events: none;
+ border-radius: inherit;
+ border: var(--n-border);
+ transition:
+ box-shadow .3s var(--n-bezier),
+ border-color .3s var(--n-bezier);
+ `),c(`state-border`,`
+ border-color: #0000;
+ z-index: 1;
+ `),c(`prefix`,`margin-right: 4px;`),c(`suffix`,`
+ margin-left: 4px;
+ `),c(`suffix, prefix`,`
+ transition: color .3s var(--n-bezier);
+ flex-wrap: nowrap;
+ flex-shrink: 0;
+ line-height: var(--n-height);
+ white-space: nowrap;
+ display: inline-flex;
+ align-items: center;
+ justify-content: center;
+ color: var(--n-suffix-text-color);
+ `,[r(`base-loading`,`
+ font-size: var(--n-icon-size);
+ margin: 0 2px;
+ color: var(--n-loading-color);
+ `),r(`base-clear`,`
+ font-size: var(--n-icon-size);
+ `,[c(`placeholder`,[r(`base-icon`,`
+ transition: color .3s var(--n-bezier);
+ color: var(--n-icon-color);
+ font-size: var(--n-icon-size);
+ `)])]),f(`>`,[r(`icon`,`
+ transition: color .3s var(--n-bezier);
+ color: var(--n-icon-color);
+ font-size: var(--n-icon-size);
+ `)]),r(`base-icon`,`
+ font-size: var(--n-icon-size);
+ `)]),r(`input-word-count`,`
+ pointer-events: none;
+ line-height: 1.5;
+ font-size: .85em;
+ color: var(--n-count-text-color);
+ transition: color .3s var(--n-bezier);
+ margin-left: 4px;
+ font-variant: tabular-nums;
+ `),[`warning`,`error`].map(e=>m(`${e}-status`,[M(`disabled`,[r(`base-loading`,`
+ color: var(--n-loading-color-${e})
+ `),c(`input-el, textarea-el`,`
+ caret-color: var(--n-caret-color-${e});
+ `),c(`state-border`,`
+ border: var(--n-border-${e});
+ `),f(`&:hover`,[c(`state-border`,`
+ border: var(--n-border-hover-${e});
+ `)]),f(`&:focus`,`
+ background-color: var(--n-color-focus-${e});
+ `,[c(`state-border`,`
+ box-shadow: var(--n-box-shadow-focus-${e});
+ border: var(--n-border-focus-${e});
+ `)]),m(`focus`,`
+ background-color: var(--n-color-focus-${e});
+ `,[c(`state-border`,`
+ box-shadow: var(--n-box-shadow-focus-${e});
+ border: var(--n-border-focus-${e});
+ `)])])]))]),pe=r(`input`,[m(`disabled`,[c(`input-el, textarea-el`,`
+ -webkit-text-fill-color: var(--n-text-color-disabled);
+ `)])]);function H(e){let t=0;for(let n of e)t++;return t}function U(e){return e===``||e==null}function me(e){let t=l(null);function n(){let{value:n}=e;if(!n?.focus){i();return}let{selectionStart:r,selectionEnd:a,value:o}=n;if(r==null||a==null){i();return}t.value={start:r,end:a,beforeText:o.slice(0,r),afterText:o.slice(a)}}function r(){let{value:n}=t,{value:r}=e;if(!n||!r)return;let{value:i}=r,{start:a,beforeText:o,afterText:s}=n,c=i.length;if(i.endsWith(s))c=i.length-s.length;else if(i.startsWith(o))c=o.length;else{let e=o[a-1],t=i.indexOf(e,a-1);t!==-1&&(c=t+1)}r.setSelectionRange?.(c,c)}function i(){t.value=null}return g(e,i),{recordCursor:n,restoreCursor:r}}var W=O({name:`InputWordCount`,setup(e,{slots:t}){let{mergedValueRef:n,maxlengthRef:r,mergedClsPrefixRef:i,countGraphemesRef:a}=A(V),o=b(()=>{let{value:e}=n;return e===null||Array.isArray(e)?0:(a.value||H)(e)});return()=>{let{value:e}=r,{value:a}=n;return j(),S(`span`,{class:v(`${i.value}-input-word-count`)},[y(()=>w(t.default,{value:a===null||Array.isArray(a)?``:a},()=>[e===void 0?o.value:`${o.value} / ${e}`]))],2)}}}),he=[`autofocus`,`rows`,`placeholder`,`value`,`disabled`,`maxlength`,`minlength`,`readonly`,`tabindex`,`onBlur`,`onFocus`,`onInput`,`onChange`,`onScroll`],ge=[`type`,`tabindex`,`placeholder`,`disabled`,`maxlength`,`minlength`,`value`,`readonly`,`autofocus`,`size`,`onBlur`,`onFocus`,`onInput`,`onChange`],G=[`onMousedown`,`onClick`],_e=[`type`,`tabindex`,`placeholder`,`disabled`,`maxlength`,`minlength`,`value`,`readonly`,`onBlur`,`onFocus`,`onInput`,`onChange`],K=[`tabindex`,`onFocus`,`onBlur`,`onClick`,`onMousedown`,`onMouseenter`,`onMouseleave`,`onCompositionstart`,`onCompositionend`,`onKeyup`,`onKeydown`],q={...a.props,bordered:{type:Boolean,default:void 0},type:{type:String,default:`text`},placeholder:[Array,String],defaultValue:{type:[String,Array],default:null},value:[String,Array],disabled:{type:Boolean,default:void 0},size:String,rows:{type:[Number,String],default:3},round:Boolean,minlength:[String,Number],maxlength:[String,Number],clearable:Boolean,autosize:{type:[Boolean,Object],default:!1},pair:Boolean,separator:String,readonly:{type:[String,Boolean],default:!1},passivelyActivated:Boolean,showPasswordOn:String,stateful:{type:Boolean,default:!0},autofocus:Boolean,inputProps:Object,resizable:{type:Boolean,default:!0},showCount:Boolean,loading:{type:Boolean,default:void 0},allowInput:Function,renderCount:Function,onMousedown:Function,onKeydown:Function,onKeyup:[Function,Array],onInput:[Function,Array],onFocus:[Function,Array],onBlur:[Function,Array],onClick:[Function,Array],onChange:[Function,Array],onClear:[Function,Array],countGraphemes:Function,status:String,"onUpdate:value":[Function,Array],onUpdateValue:[Function,Array],textDecoration:[String,Array],attrSize:{type:Number,default:20},onInputBlur:[Function,Array],onInputFocus:[Function,Array],onDeactivate:[Function,Array],onActivate:[Function,Array],onWrapperFocus:[Function,Array],onWrapperBlur:[Function,Array],internalDeactivateOnEnter:Boolean,internalForceFocus:Boolean,internalLoadingBeforeSuffix:{type:Boolean,default:!0},showPasswordToggle:Boolean},ve=O({name:`Input`,props:q,slots:Object,setup(e){let{mergedClsPrefixRef:r,mergedBorderedRef:o,inlineThemeDisabled:c,mergedRtlRef:d,mergedComponentPropsRef:f}=oe(e),p=a(`Input`,`-input`,fe,de,e,r);ae&&te(`-input-safari`,pe,r);let m=l(null),h=l(null),_=l(null),v=l(null),y=l(null),x=l(null),S=l(null),w=me(S),T=l(null),{localeRef:D}=le(`Input`),O=l(e.defaultValue),k=i(e,`value`),A=ue(k,O),j=ce(e,{mergedSize:t=>{let{size:n}=e;if(n)return n;let{mergedSize:r}=t||{};return r?.value?r.value:f?.value?.Input?.size||`medium`}}),{mergedSizeRef:M,mergedDisabledRef:N,mergedStatusRef:P}=j,I=l(!1),L=l(!1),z=l(!1),B=l(!1),H=null,W=b(()=>{let{placeholder:t,pair:n}=e;return n?Array.isArray(t)?t:t===void 0?[``,``]:[t,t]:t===void 0?[D.value.placeholder]:[t]}),he=b(()=>{let{value:e}=z,{value:t}=A,{value:n}=W;return!e&&(U(t)||Array.isArray(t)&&U(t[0]))&&n[0]}),ge=b(()=>{let{value:e}=z,{value:t}=A,{value:n}=W;return!e&&n[1]&&(U(t)||Array.isArray(t)&&U(t[1]))}),G=u(()=>e.internalForceFocus||I.value),_e=u(()=>{if(N.value||e.readonly||!e.clearable||!G.value&&!L.value)return!1;let{value:t}=A,{value:n}=G;return e.pair?!!(Array.isArray(t)&&(t[0]||t[1]))&&(L.value||n):!!t&&(L.value||n)}),K=b(()=>{let{showPasswordOn:t}=e;if(t)return t;if(e.showPasswordToggle)return`click`}),q=l(!1),ve=b(()=>{let{textDecoration:t}=e;return t?Array.isArray(t)?t.map(e=>({textDecoration:e})):[{textDecoration:t}]:[``,``]}),ye=l(void 0),be=()=>{if(e.type===`textarea`){let{autosize:t}=e;if(t&&(ye.value=T.value?.$el?.offsetWidth),!h.value||typeof t==`boolean`)return;let{paddingTop:n,paddingBottom:r,lineHeight:i}=window.getComputedStyle(h.value),a=Number(n.slice(0,-2)),o=Number(r.slice(0,-2)),s=Number(i.slice(0,-2)),{value:c}=_;if(!c)return;if(t.minRows){let e=Math.max(t.minRows,1),n=`${a+o+s*e}px`;c.style.minHeight=n}if(t.maxRows){let e=`${a+o+s*t.maxRows}px`;c.style.maxHeight=e}}},xe=b(()=>{let{maxlength:t}=e;return t===void 0?void 0:Number(t)});re(()=>{let{value:e}=A;Array.isArray(e)||$(e)});let Se=ie().proxy;function J(t,n){let{onUpdateValue:r,"onUpdate:value":i,onInput:a}=e,{nTriggerFormInput:o}=j;r&&C(r,t,n),i&&C(i,t,n),a&&C(a,t,n),O.value=t,o()}function Y(t,n){let{onChange:r}=e,{nTriggerFormChange:i}=j;r&&C(r,t,n),O.value=t,i()}function Ce(t){let{onBlur:n}=e,{nTriggerFormBlur:r}=j;n&&C(n,t),r()}function we(t){let{onFocus:n}=e,{nTriggerFormFocus:r}=j;n&&C(n,t),r()}function Te(t){let{onClear:n}=e;n&&C(n,t)}function Ee(t){let{onInputBlur:n}=e;n&&C(n,t)}function De(t){let{onInputFocus:n}=e;n&&C(n,t)}function Oe(){let{onDeactivate:t}=e;t&&C(t)}function ke(){let{onActivate:t}=e;t&&C(t)}function Ae(t){let{onClick:n}=e;n&&C(n,t)}function je(t){let{onWrapperFocus:n}=e;n&&C(n,t)}function Me(t){let{onWrapperBlur:n}=e;n&&C(n,t)}function Ne(){z.value=!0}function Pe(e){z.value=!1,e.target===x.value?X(e,1):X(e,0)}function X(t,n=0,r=`input`){let i=t.target.value;if($(i),t instanceof InputEvent&&!t.isComposing&&(z.value=!1),e.type===`textarea`){let{value:e}=T;e&&e.syncUnifiedContainer()}if(H=i,z.value)return;w.recordCursor();let a=Fe(i);if(a){if(!e.pair)r===`input`?J(i,{source:n}):Y(i,{source:n});else{let{value:e}=A;e=Array.isArray(e)?[e[0],e[1]]:[``,``],e[n]=i,r===`input`?J(e,{source:n}):Y(e,{source:n})}}Se.$forceUpdate(),a||E(w.restoreCursor)}function Fe(t){let{countGraphemes:n,maxlength:r,minlength:i}=e;if(n){let e;if(r!==void 0&&(e===void 0&&(e=n(t)),e>Number(r))||i!==void 0&&(e===void 0&&(e=n(t)),e<Number(r)))return!1}let{allowInput:a}=e;return typeof a!=`function`||a(t)}function Ie(e){Ee(e),e.relatedTarget===m.value&&Oe(),(e.relatedTarget===null||e.relatedTarget!==y.value&&e.relatedTarget!==x.value&&e.relatedTarget!==h.value)&&(B.value=!1),Z(e,`blur`),S.value=null}function Le(e,t){De(e),I.value=!0,B.value=!0,ke(),Z(e,`focus`),t===0?S.value=y.value:t===1?S.value=x.value:t===2&&(S.value=h.value)}function Re(t){e.passivelyActivated&&(Me(t),Z(t,`blur`))}function ze(t){e.passivelyActivated&&(I.value=!0,je(t),Z(t,`focus`))}function Z(e,t){e.relatedTarget!==null&&(e.relatedTarget===y.value||e.relatedTarget===x.value||e.relatedTarget===h.value||e.relatedTarget===m.value)||(t===`focus`?(we(e),I.value=!0):t===`blur`&&(Ce(e),I.value=!1))}function Be(e,t){X(e,t,`change`)}function Ve(e){Ae(e)}function He(e){Te(e),Ue()}function Ue(){e.pair?(J([``,``],{source:`clear`}),Y([``,``],{source:`clear`})):(J(``,{source:`clear`}),Y(``,{source:`clear`}))}function We(t){let{onMousedown:n}=e;n&&n(t);let{tagName:r}=t.target;if(r!==`INPUT`&&r!==`TEXTAREA`){if(e.resizable){let{value:e}=m;if(e){let{left:n,top:r,width:i,height:a}=e.getBoundingClientRect();if(n+i-14<t.clientX&&t.clientX<n+i&&r+a-14<t.clientY&&t.clientY<r+a)return}}t.preventDefault(),I.value||Qe()}}function Ge(){L.value=!0,e.type===`textarea`&&T.value?.handleMouseEnterWrapper()}function Ke(){L.value=!1,e.type===`textarea`&&T.value?.handleMouseLeaveWrapper()}function qe(){N.value||K.value===`click`&&(q.value=!q.value)}function Je(e){if(N.value)return;e.preventDefault();let t=e=>{e.preventDefault(),R(`mouseup`,document,t)};if(F(`mouseup`,document,t),K.value!==`mousedown`)return;q.value=!0;let n=()=>{q.value=!1,R(`mouseup`,document,n)};F(`mouseup`,document,n)}function Ye(t){e.onKeyup&&C(e.onKeyup,t)}function Xe(t){switch(e.onKeydown&&C(e.onKeydown,t),t.key){case`Escape`:Q();break;case`Enter`:Ze(t)}}function Ze(t){if(e.passivelyActivated){let{value:n}=B;if(n){e.internalDeactivateOnEnter&&Q();return}t.preventDefault(),e.type===`textarea`?h.value?.focus():y.value?.focus()}}function Q(){e.passivelyActivated&&(B.value=!1,E(()=>{m.value?.focus()}))}function Qe(){N.value||(e.passivelyActivated?m.value?.focus():(h.value?.focus(),y.value?.focus()))}function $e(){m.value?.contains(document.activeElement)&&document.activeElement.blur()}function et(){h.value?.select(),y.value?.select()}function tt(){N.value||(h.value?h.value.focus():y.value&&y.value.focus())}function nt(){let{value:e}=m;e?.contains(document.activeElement)&&e!==document.activeElement&&Q()}function rt(t){if(e.type===`textarea`){let{value:e}=h;e?.scrollTo(t)}else{let{value:e}=y;e?.scrollTo(t)}}function $(t){let{type:n,pair:r,autosize:i}=e;if(!r&&i){if(n===`textarea`){let{value:e}=_;e&&(e.textContent=`${t??``}\r\n`)}else{let{value:e}=v;e&&(t?e.textContent=t:e.innerHTML=`&nbsp;`)}}}function it(){be()}let at=l({top:`0`});function ot(e){let{scrollTop:t}=e.target;at.value.top=`${-t}px`,T.value?.syncUnifiedContainer()}let st=null;n(()=>{let{autosize:t,type:n}=e;t&&n===`textarea`?st=g(A,e=>{!Array.isArray(e)&&e!==H&&$(e)}):st?.()});let ct=null;n(()=>{e.type===`textarea`?ct=g(A,e=>{!Array.isArray(e)&&e!==H&&T.value?.syncUnifiedContainer()}):ct?.()}),se(V,{mergedValueRef:A,maxlengthRef:xe,mergedClsPrefixRef:r,countGraphemesRef:i(e,`countGraphemes`)});let lt={wrapperElRef:m,inputElRef:y,textareaElRef:h,isCompositing:z,clear:Ue,focus:Qe,blur:$e,select:et,deactivate:nt,activate:tt,scrollTo:rt},ut=ne(`Input`,d,r),dt=b(()=>{let{value:e}=M,{common:{cubicBezierEaseInOut:n},self:{color:r,colorHover:i,borderRadius:a,textColor:o,caretColor:c,caretColorError:ee,caretColorWarning:l,textDecorationColor:u,border:d,borderDisabled:f,borderHover:te,borderFocus:m,placeholderColor:h,placeholderColorDisabled:g,lineHeightTextarea:ne,colorDisabled:_,colorFocus:v,textColorDisabled:y,boxShadowFocus:re,iconSize:ie,colorFocusWarning:b,boxShadowFocusWarning:x,borderWarning:ae,borderFocusWarning:S,borderHoverWarning:C,colorFocusError:w,boxShadowFocusError:T,borderError:E,borderFocusError:D,borderHoverError:O,clearSize:oe,clearColor:k,clearColorHover:A,clearColorPressed:j,iconColor:se,iconColorDisabled:ce,suffixTextColor:le,countTextColor:N,countTextColorDisabled:ue,iconColorHover:P,iconColorPressed:F,loadingColor:I,loadingColorError:L,loadingColorWarning:R,fontWeight:de,[t(`padding`,e)]:z,[t(`fontSize`,e)]:B,[t(`height`,e)]:V}}=p.value,{left:fe,right:pe}=s(z);return{"--n-bezier":n,"--n-count-text-color":N,"--n-count-text-color-disabled":ue,"--n-color":r,"--n-color-hover":i,"--n-font-size":B,"--n-font-weight":de,"--n-border-radius":a,"--n-height":V,"--n-padding-left":fe,"--n-padding-right":pe,"--n-text-color":o,"--n-caret-color":c,"--n-text-decoration-color":u,"--n-border":d,"--n-border-disabled":f,"--n-border-hover":te,"--n-border-focus":m,"--n-placeholder-color":h,"--n-placeholder-color-disabled":g,"--n-icon-size":ie,"--n-line-height-textarea":ne,"--n-color-disabled":_,"--n-color-focus":v,"--n-text-color-disabled":y,"--n-box-shadow-focus":re,"--n-loading-color":I,"--n-caret-color-warning":l,"--n-color-focus-warning":b,"--n-box-shadow-focus-warning":x,"--n-border-warning":ae,"--n-border-focus-warning":S,"--n-border-hover-warning":C,"--n-loading-color-warning":R,"--n-caret-color-error":ee,"--n-color-focus-error":w,"--n-box-shadow-focus-error":T,"--n-border-error":E,"--n-border-focus-error":D,"--n-border-hover-error":O,"--n-loading-color-error":L,"--n-clear-color":k,"--n-clear-size":oe,"--n-clear-color-hover":A,"--n-clear-color-pressed":j,"--n-icon-color":se,"--n-icon-color-hover":P,"--n-icon-color-pressed":F,"--n-icon-color-disabled":ce,"--n-suffix-text-color":le}}),ft=c?ee(`input`,b(()=>{let{value:e}=M;return e[0]}),dt,e):void 0;return{...lt,wrapperElRef:m,inputElRef:y,inputMirrorElRef:v,inputEl2Ref:x,textareaElRef:h,textareaMirrorElRef:_,textareaScrollbarInstRef:T,rtlEnabled:ut,uncontrolledValue:O,mergedValue:A,passwordVisible:q,mergedPlaceholder:W,showPlaceholder1:he,showPlaceholder2:ge,mergedFocus:G,isComposing:z,activated:B,showClearButton:_e,mergedSize:M,mergedDisabled:N,textDecorationStyle:ve,mergedClsPrefix:r,mergedBordered:o,mergedShowPasswordOn:K,placeholderStyle:at,mergedStatus:P,textAreaScrollContainerWidth:ye,handleTextAreaScroll:ot,handleCompositionStart:Ne,handleCompositionEnd:Pe,handleInput:X,handleInputBlur:Ie,handleInputFocus:Le,handleWrapperBlur:Re,handleWrapperFocus:ze,handleMouseEnter:Ge,handleMouseLeave:Ke,handleMouseDown:We,handleChange:Be,handleClick:Ve,handleClear:He,handlePasswordToggleClick:qe,handlePasswordToggleMousedown:Je,handleWrapperKeydown:Xe,handleWrapperKeyup:Ye,handleTextAreaMirrorResize:it,getTextareaScrollContainer:()=>h.value,mergedTheme:p,cssVars:c?void 0:dt,themeClass:ft?.themeClass,onRender:ft?.onRender}},render(){let{mergedClsPrefix:t,mergedStatus:n,themeClass:r,type:i,countGraphemes:a,onRender:o}=this,s=this.$slots;return o?.(),j(),S(`div`,{ref:`wrapperElRef`,class:v([`${t}-input`,`${t}-input--${this.mergedSize}-size`,r,n&&`${t}-input--${n}-status`,{[`${t}-input--rtl`]:this.rtlEnabled,[`${t}-input--disabled`]:this.mergedDisabled,[`${t}-input--textarea`]:i===`textarea`,[`${t}-input--resizable`]:this.resizable&&!this.autosize,[`${t}-input--autosize`]:this.autosize,[`${t}-input--round`]:this.round&&i!==`textarea`,[`${t}-input--pair`]:this.pair,[`${t}-input--focus`]:this.mergedFocus,[`${t}-input--stateful`]:this.stateful}]),style:h(this.cssVars),tabindex:!this.mergedDisabled&&this.passivelyActivated&&!this.activated?0:void 0,onFocus:this.handleWrapperFocus,onBlur:this.handleWrapperBlur,onClick:this.handleClick,onMousedown:this.handleMouseDown,onMouseenter:this.handleMouseEnter,onMouseleave:this.handleMouseLeave,onCompositionstart:this.handleCompositionStart,onCompositionend:this.handleCompositionEnd,onKeyup:this.handleWrapperKeyup,onKeydown:this.handleWrapperKeydown},[k(`div`,{class:v(`${t}-input-wrapper`)},[y(()=>e(s.prefix,e=>e&&(j(),S(`div`,{class:v(`${t}-input__prefix`)},[y(()=>e)],2)))),i===`textarea`?(j(),T(L,{key:0,ref:`textareaScrollbarInstRef`,class:v(`${t}-input__textarea`),container:this.getTextareaScrollContainer,theme:this.theme?.peers?.Scrollbar,themeOverrides:this.themeOverrides?.peers?.Scrollbar,triggerDisplayManually:!0,useUnifiedContainer:!0,internalHoistYRail:!0},{default:()=>{let{textAreaScrollContainerWidth:e}=this,n={width:this.autosize&&e&&`${e}px`};return j(),S(_,null,[k(`textarea`,x(this.inputProps,{ref:`textareaElRef`,class:[`${t}-input__textarea-el`,this.inputProps?.class],autofocus:this.autofocus,rows:Number(this.rows),placeholder:this.placeholder,value:this.mergedValue,disabled:this.mergedDisabled,maxlength:a?void 0:this.maxlength,minlength:a?void 0:this.minlength,readonly:this.readonly,tabindex:this.passivelyActivated&&!this.activated?-1:void 0,style:[this.textDecorationStyle[0],this.inputProps?.style,n],onBlur:this.handleInputBlur,onFocus:e=>{this.handleInputFocus(e,2)},onInput:this.handleInput,onChange:this.handleChange,onScroll:this.handleTextAreaScroll}),null,16,he),this.showPlaceholder1?(j(),S(`div`,{class:v(`${t}-input__placeholder`),style:h([this.placeholderStyle,n]),key:`placeholder`},[y(()=>this.mergedPlaceholder[0])],6)):y(()=>null),this.autosize?(j(),T(I,{key:2,onResize:this.handleTextAreaMirrorResize},{default:()=>(j(),S(`div`,{ref:`textareaMirrorElRef`,class:v(`${t}-input__textarea-mirror`),key:`mirror`},null,2))},1032,[`onResize`])):y(()=>null)],64)}},1032,[`class`,`container`,`theme`,`themeOverrides`])):(j(),S(`div`,{key:1,class:v(`${t}-input__input`)},[k(`input`,x({type:i===`password`&&this.mergedShowPasswordOn&&this.passwordVisible?`text`:i},this.inputProps,{ref:`inputElRef`,class:[`${t}-input__input-el`,this.inputProps?.class],style:[this.textDecorationStyle[0],this.inputProps?.style],tabindex:this.passivelyActivated&&!this.activated?-1:this.inputProps?.tabindex,placeholder:this.mergedPlaceholder[0],disabled:this.mergedDisabled,maxlength:a?void 0:this.maxlength,minlength:a?void 0:this.minlength,value:Array.isArray(this.mergedValue)?this.mergedValue[0]:this.mergedValue,readonly:this.readonly,autofocus:this.autofocus,size:this.attrSize,onBlur:this.handleInputBlur,onFocus:e=>{this.handleInputFocus(e,0)},onInput:e=>{this.handleInput(e,0)},onChange:e=>{this.handleChange(e,0)}}),null,16,ge),this.showPlaceholder1?(j(),S(`div`,{key:0,class:v(`${t}-input__placeholder`)},[k(`span`,null,[y(()=>this.mergedPlaceholder[0])])],2)):y(()=>null),this.autosize?(j(),S(`div`,{class:v(`${t}-input__input-mirror`),key:`mirror`,ref:`inputMirrorElRef`},`\xA0`,2)):y(()=>null)],2)),y(()=>!this.pair&&e(s.suffix,n=>n||this.clearable||this.showCount||this.mergedShowPasswordOn||this.loading!==void 0?(j(),S(`div`,{key:1,class:v(`${t}-input__suffix`)},[y(()=>[e(s[`clear-icon-placeholder`],e=>(this.clearable||e)&&(j(),T(N,{clsPrefix:t,show:this.showClearButton,onClear:this.handleClear},{placeholder:()=>e,icon:()=>this.$slots[`clear-icon`]?.()},1032,[`clsPrefix`,`show`,`onClear`]))),this.internalLoadingBeforeSuffix?null:n,this.loading===void 0?null:(j(),T(P,{key:2,clsPrefix:t,loading:this.loading,showArrow:!1,showClear:!1,style:h(this.cssVars)},null,8,[`clsPrefix`,`loading`,`style`])),this.internalLoadingBeforeSuffix?n:null,this.showCount&&this.type!==`textarea`?(j(),T(W,{key:3},{default:e=>{let{renderCount:t}=this;return t?t(e):s.count?.(e)}},1024)):null,this.mergedShowPasswordOn&&this.type===`password`?(j(),S(`div`,{key:4,class:v(`${t}-input__eye`),onMousedown:this.handlePasswordToggleMousedown,onClick:this.handlePasswordToggleClick},[this.passwordVisible?(j(),S(_,{key:0},[y(()=>d(s[`password-visible-icon`],()=>[(j(),T(p,{clsPrefix:t},{default:()=>(j(),T(z))},1032,[`clsPrefix`]))]))],64)):(j(),S(_,{key:1},[y(()=>d(s[`password-invisible-icon`],()=>[(j(),T(p,{clsPrefix:t},{default:()=>(j(),T(B))},1032,[`clsPrefix`]))]))],64))],42,G)):null])],2)):null))],2),this.pair?(j(),S(`span`,{key:0,class:v(`${t}-input__separator`)},[y(()=>d(s.separator,()=>[this.separator]))],2)):y(()=>null),this.pair?(j(),S(`div`,{key:2,class:v(`${t}-input-wrapper`)},[k(`div`,{class:v(`${t}-input__input`)},[k(`input`,{ref:`inputEl2Ref`,type:this.type,class:v(`${t}-input__input-el`),tabindex:this.passivelyActivated&&!this.activated?-1:void 0,placeholder:this.mergedPlaceholder[1],disabled:this.mergedDisabled,maxlength:a?void 0:this.maxlength,minlength:a?void 0:this.minlength,value:Array.isArray(this.mergedValue)?this.mergedValue[1]:void 0,readonly:this.readonly,style:h(this.textDecorationStyle[1]),onBlur:this.handleInputBlur,onFocus:e=>{this.handleInputFocus(e,1)},onInput:e=>{this.handleInput(e,1)},onChange:e=>{this.handleChange(e,1)}},null,46,_e),this.showPlaceholder2?(j(),S(`div`,{key:0,class:v(`${t}-input__placeholder`)},[k(`span`,null,[y(()=>this.mergedPlaceholder[1])])],2)):y(()=>null)],2),y(()=>e(s.suffix,e=>(this.clearable||e)&&(j(),S(`div`,{class:v(`${t}-input__suffix`)},[y(()=>[this.clearable&&(j(),T(N,{clsPrefix:t,show:this.showClearButton,onClear:this.handleClear},{icon:()=>s[`clear-icon`]?.(),placeholder:()=>s[`clear-icon-placeholder`]?.()},1032,[`clsPrefix`,`show`,`onClear`])),e])],2))))],2)):y(()=>null),this.mergedBordered?(j(),S(`div`,{key:4,class:v(`${t}-input__border`)},null,2)):y(()=>null),this.mergedBordered?(j(),S(`div`,{key:6,class:v(`${t}-input__state-border`)},null,2)):y(()=>null),this.showCount&&i===`textarea`?(j(),T(W,{key:8},{default:e=>{let{renderCount:t}=this;return t?t(e):s.count?.(e)}},1024)):y(()=>null)],46,K)}});export{ve as t};